refactor: moved setBoxLabels to labelUtil
