+++ src/chart/boxplot/BoxplotView.ts
@@ -29,6 +29,7 @@ import SeriesData from '../../data/SeriesData';
 import { BoxplotItemLayout } from './boxplotLayout';
 import { saveOldStyle } from '../../animation/basicTransition';
 import { ViewRootGroup } from '../../util/types';
+import { setBoxLabels } from '../../label/labelStyle';
 
 class BoxplotView extends ChartView {
     static type = 'boxplot';
@@ -187,70 +188,19 @@ function updateNormalBoxData(
 
     const itemModel = data.getItemModel<BoxplotDataItemOption>(dataIndex);
     const emphasisModel = itemModel.getModel('emphasis');
+    const labelModel = seriesModel.get('label');
 
-    if (seriesModel.option.label.show) {
+    if (labelModel.show) {
         const formattedLabels =
             ((seriesModel as BoxplotSeriesModel).getRawValue(dataIndex) as number[])
-            .splice(1).map((value: number) => seriesModel.option.label.formatter(value).toString());
+            .splice(1).map((value: number) => labelModel.formatter(value).toString());
 
-        setBoxLabels(itemLayout.ends, formattedLabels, seriesModel.option.label, group);
+        setBoxLabels(itemLayout.ends, formattedLabels, labelModel, group);
     }
 
     setStatesStylesFromModel(el, itemModel);
 
     toggleHoverEmphasis(el, emphasisModel.get('focus'), emphasisModel.get('blurScope'), emphasisModel.get('disabled'));
-}
-
-function setBoxLabels(boxItemPositions: number[][], formattedLabels: Array<string>,
-    labelOption: any, group: ViewRootGroup) {
-    // get sorted and unique y positions of box items
-    const yPositions = boxItemPositions.map((pos: number[]) => pos[1]);
-    const uniqueYPositions: number[] = [];
-    yPositions.forEach(position => {
-        if (!uniqueYPositions.includes(position)) {
-            uniqueYPositions.push(position);
-        }
-    });
-    uniqueYPositions.sort(function (a: number, b: number) {
-        return b - a;
-    });
-
-    boxItemPositions.sort(function (a: number[], b: number[]) {
-        return a[0] - b[0];
-    });
-
-    // get alternating y positions for labels to avoid overlap
-    const uniqueAlternatingPositions = uniqueYPositions.map((posY: number, ind: number) => {
-            const matchingPositions = boxItemPositions.filter((orgPos: number[]) => orgPos[1] === posY);
-            const index = (ind % 2 === 0) ? 0 : matchingPositions.length - 1;
-            return matchingPositions[index];
-    });
-
-    // create labels and add them to their respective positions
-    formattedLabels.forEach((labelText: string, ind: number) => {
-        if (labelOption.show === 'iqr' && (ind === 0 || ind === 4)) {
-            return;
-        }
-        if (labelOption.show === 'median' && (ind !== 2)) {
-            return;
-        }
-        if (labelOption.show === 'whiskers' && (ind !== 0 && ind !== 4)) {
-            return;
-        }
-        const label = new graphic.Text({
-            style: {
-                text: labelText
-            },
-            z2: 1000
-        });
-        const defaultOffset = 5;
-        const defaultXOffset = (ind % 2) === 0 ? (-(label.getBoundingRect().width + defaultOffset)) : defaultOffset;
-        const defaultYOffset = -defaultOffset;
-        const customOffset = labelOption.offset ? labelOption.offset : [0, 0];
-        label.x = uniqueAlternatingPositions[ind][0] + defaultXOffset + customOffset[0],
-        label.y = uniqueAlternatingPositions[ind][1] + defaultYOffset + customOffset[1],
-        group.add(label);
-    });
 }
 
 function transInit(points: number[][], dim: number, itemLayout: BoxplotItemLayout) {
